fix: Resolve compilation errors and add testing rules
Critical Fixes:
- Add missing FEE import at file level in utxo_manager.rs
- Fix type mismatch: usize to u64 conversion in split_tx
- Resolve all compilation errors for mass limit solution

New Development Rule:
- NEVER PUSH WITHOUT TESTING - absolute rule added to CLAUDE.md
- Must confirm cargo build succeeds before any git push
- Prevents broken code from reaching repository

+++ examples/comment-board/src/wallet/utxo_manager.rs
@@ -7,6 +7,7 @@ use log::*;
 
 // Phase 2.0: Import script generation for true UTXO locking
 use crate::wallet::kaspa_scripts::{ScriptUnlockCondition, create_bond_script_pubkey, validate_script_conditions};
+use crate::utils::FEE;
 
 /// Real UTXO Locking Manager for Economic Episode Contracts - Phase 1.1 Implementation
 #[derive(Debug, Clone)]
@@ -515,7 +516,7 @@ impl UtxoLockManager {
             let split_tx = generator.build_transaction(
                 &utxos_to_use,
                 chunk_size * num_outputs as u64, // Total amount minus fee
-                num_outputs,
+                num_outputs as u64, // Convert usize to u64
                 &self.kaspa_address,
                 "UTXO_SPLIT".as_bytes().to_vec(), // Minimal payload
             );
